synapse-prometheus-connector/src/requirements.txt
requests==2.24.0
python-dateutil==2.8.1
attrs==19.3.0
cattrs==1.0.0
PyYAML==5.3.1
prometheus-client==0.8.0
timestring==1.6.4